+++ resources/assets/js/bootstrap.js
@@ -37,4 +37,3 @@ require('./modules/post');
 require('./modules/datetimepicker');
 require('./modules/leftMenu');
 require('./modules/open-click');
-require('./modules/shortcut');