+++ resources/stubs/database/migrations/2015_10_19_214425_create_orchid_role_users_table.php
@@ -22,6 +22,6 @@ class CreateOrchidRoleUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('role_users');
+        Schema::dropIfExists('role_users');
     }
 }