+++ resources/stubs/database/migrations/2015_10_19_214425_create_orchid_roles_table.php
@@ -24,6 +24,6 @@ class CreateOrchidRolesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('roles');
+        Schema::dropIfExists('roles');
     }
 }